requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.1.0
langchain-chroma>=0.1.0
langchain-experimental>=0.1.0
chromadb>=0.4.0
faiss-cpu>=1.7.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
Pillow>=10.0.0
torch>=2.0.0
transformers>=4.30.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0 